mergeChampionsData
